+++ src/pop/nodevector.h
@@ -47,20 +47,20 @@ namespace FT{
 
             size_t subtree(size_t i, char otype='0') const;
             
-            void set_weights(vector<vector<double>>& weights);
+            void set_weights(vector<vector<float>>& weights);
             
-            vector<vector<double>> get_weights();
+            vector<vector<float>> get_weights();
             
             bool is_valid_program(unsigned num_features, 
                                   vector<string> longitudinalMap);
        
             void make_tree(const NodeVector& functions, 
                            const NodeVector& terminals, int max_d,  
-                           const vector<double>& term_weights, char otype, const vector<char>& term_types);
+                           const vector<float>& term_weights, char otype, const vector<char>& term_types);
 
             void make_program(const NodeVector& functions, 
                               const NodeVector& terminals, int max_d, 
-                              const vector<double>& term_weights, int dim, char otype, 
+                              const vector<float>& term_weights, int dim, char otype, 
                               vector<string> longitudinalMap, const vector<char>& term_types);
             
         }; //NodeVector
