+++ src/pop/op/arithmetic/n_log.h
@@ -15,15 +15,15 @@ namespace FT{
             {
             	public:
             	
-            		NodeLog(vector<double> W0 = vector<double>());
+            		NodeLog(vector<float> W0 = vector<float>());
             		
-                    /// Safe log: pushes log(abs(x)) or MIN_DBL if x is near zero. 
+                    /// Safe log: pushes log(abs(x)) or MIN_FLT if x is near zero. 
                     void evaluate(const Data& data, State& state);
 
                     /// Evaluates the node symbolically
                     void eval_eqn(State& state);
 
-                    ArrayXd getDerivative(Trace& state, int loc);
+                    ArrayXf getDerivative(Trace& state, int loc);
                     
                 protected:
                     NodeLog* clone_impl() const override;
